+++ packages/components/src/Button/BaseButton.vue
@@ -1,8 +1,3 @@
-<script lang="ts">
-export const variants = ["filled", "outlined", "transparent"] as const;
-export type Variant = (typeof variants)[number];
-</script>
-
 <script setup lang="ts" generic="UNUSED">
 import { type Component, computed } from "vue";
 
@@ -98,7 +93,7 @@ html.kds-legacy {
 
   &:focus-visible {
     outline: var(--kds-border-action-focused);
-    outline-offset: var(--kds-focus-outline-offset);
+    outline-offset: var(--kds-spacing-offset-focus);
   }
 
   &.filled {