+++ packages/components/src/Button/Button.vue
@@ -1,8 +1,7 @@
 <script setup lang="ts" generic="UNUSED">
 import BaseButton from "./BaseButton.vue";
-import type { BaseButtonProps } from "./types";
+import type { ButtonProps } from "./types";
 
-export type ButtonProps = BaseButtonProps & {}; // intersection needed for Storybook to work
 const props = defineProps<ButtonProps>();
 
 const emit = defineEmits<{