+++ packages/components/src/Icon/types.ts
@@ -1,5 +1,9 @@
-import { sizes } from "../constants";
-import type { Size } from "../types";
+import type { IconName } from "@knime/kds-styles/img/icons/def";
+import type { TypeIconName } from "@knime/kds-styles/img/type-icons/def";
 
-export const iconSizes = sizes;
-export type IconSize = Size;
+import type { dataTypeSizes, iconSizes } from "./constants";
+
+type IconSize = (typeof iconSizes)[number];
+type DataTypeSize = (typeof dataTypeSizes)[number];
+
+export type { IconSize, IconName, DataTypeSize, TypeIconName };