+++ packages/styles/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @knime/kds-styles
 
+## 0.0.16
+
+### Patch Changes
+
+- f43d7c8: added size.container.width.25x, 32x, 45x
+
+## 0.0.15
+
+### Patch Changes
+
+- 6273d5e: added kds.spacing.offset.focus token
+
+## 0.0.14
+
+### Patch Changes
+
+- 8469844: added disabled.primary, disabled.danger to legacy theme
+
+## 0.0.13
+
+### Patch Changes
+
+- a277ce1: KDS-590: Add and update icons
+
+## 0.0.12
+
+### Patch Changes
+
+- 3c052ec: KDS-619: Change transparent token to also have a width
+
+## 0.0.11
+
+### Patch Changes
+
+- 9575244: KDS-577: Adding legacy theme to build
+
 ## 0.0.10
 
 ### Patch Changes